Factor out common code from lilypond_fragment and lilypond_file
I'm sure that can be brought even further, but I don't fully understand
the implications of the difference with fullpage or not.
Maybe there's no reason not to give the fullpage option to encoded
fragments as well. (what about a fragment that generates five exercises by repeating a permutated pattern? This could very well be included inline and span multiple pages)