'margin' parameter for ly and \lily (WIP)
Deals with #23

This is a first step towards implementing proper alignment with protrusion.
Actually I would prefer being able to pass the margin with arbitrary units
(instead of assuming points). This would also make it possible to calculate
a paper width from line-width plus margin, so the score is guaranteed not
to be cut off on any side.