Partially restore fragment compilation code
Unfortunately it didn't work as expected to simply switch on the
EPS backend (as claimed in commit 39163ba6). It turns out that this
produces only one single PDF instead of individual systems.

However, it's sufficient to include lilypond-book-preamble.ly
as this is guaranteed to be in the include path.

The cropping data is different too with this approach, so the calculation
had to be updated accordingly. Actually it's much simpler that way
as the first BoundingBox parameter actually gives the extent directly.