Pass current font to LilyPond (WIP)
#32

This is a first step for the process. This commit determines the font family
that is *currently* used for typesetting and passes it to all three font
families in LilyPond. This is because I can't see yet how to retrieve the
fonts for the three families in LaTeX.

I'm not completely sure about how *safe* this is. The solution I got from
http://tug.org/pipermail/luatex/2018-January/006717.html
uses several additional if tests.
I'm not sure if my code fails if the user specifies fonts differently than I
can imagine. OTOH, what should we do if the code fails to retrieve a
usable font?