New approach to determine fonts (WIP)
This is an approach suggested by Will Robertsion (maintainer/author of
fontspec): select all three fonts in turn (without printing anything) and
retrieve them one by one as the "current" font.
This *does* work - when done from the .tex document, but not when
executed from Lua.
see the following code:

\begingroup%
  \rmfamily%
  \directlua{
  myfonts = {}
    myfonts.rm = fonts.hashes.identifiers[font.current()].shared.rawdata.metadata['familyname']
  }%
  \sffamily%
  \directlua{
    myfonts.sf = fonts.hashes.identifiers[font.current()].shared.rawdata.metadata['familyname']
  }%
  \ttfamily%
  \directlua{
    myfonts.tt = fonts.hashes.identifiers[font.current()].shared.rawdata.metadata['familyname']
    print("")
    print('Main font: '..myfonts.rm)
    print('Sans font: '..myfonts.sf)
    print('Mono font: '..myfonts.tt)
  }%
\endgroup%