Pass three main font families to LilyPond
This is an approach suggested by Will Robertsion (maintainer/author of
fontspec): select all three fonts in turn (without printing anything) and
retrieve them one by one as the "current" font.

The crucial trick using \edef came from Hans Hagen
https://www.mail-archive.com/luatex@tug.org/msg06585.html

NOTE: I store the fonts in OPTIONS but *use* them with
get_local_option, which is on purpose: This will give us the possibility
to override the 'rmfamily' with the *current* font if the user wishes
so. (=> when the text currently is in \sffamily it may be desirable
to have the sans font as the main (roman) font in LilyPond too).