Better handling of boolean arguments WIP 1
Since we actually treat arguments as string arguments
we need a better way of type checking them. So far they
are only checked to equal "true" or not - which isn't actually
a proper boolean conditional.

This commit introduces a boolean check and treats
these option as actual booleans.

However, there is a severe limitation: it only works
with package options. Both local options and the
switching command don't work properly.