Avoid redundant specification of local options table
Now that we process the local options in the central function
process_local_options() we can avoid the redundancy of listing the
local options for all three situations.

While not all options are applicable to package, local and "from-here"
the list of options that apply for local option should not change between
the three places.