Handle odd/even pages differently
The pages came out in the opposite direction than desired because
we're not interested in the modulo of PAGE but the *next* page from
there.
But rather than hijacking the inner and outer margins it is better to
pass the proper page as 'first-page-number' to LilyPond anyway so
it has a chance to properly use that value.

However, it is still to be discussed how to deal with pagination in general
In musicexamples page numbers are overlaid by LaTeX and not printed
by LilyPond. We can't do that right now since we use a plain \includepdf
(and should not mess with that but leave it for the musicexamples package).

I'd say currently it's the user's responsibility to format the pagination in
LilyPond to match that of the surrounding document - or use the musicexamples
package.

Another issue is that the top and bottom margins (especially the top one) are not fully satisfactory because in the LilyPond score it is possible
that the actual score will start further down the page. But I'd say we
can discuss that in a future issue after merging.