Prepare fullpagealign option (WIP)
This commit introduces the fullpagealign option and the corresponding
command \lilypondFullpageAlign.
This option can take one out of
- crop
- staffline

The crop value simply converts TeX's current margins to LilyPond margins.
The staffline option will (does not in this commit) align the staff lines
to the type area and cause vertical protrusion if necessary.

#29 discusses which behaviour should be default, this commit sets
'crop' as default (as 'staffline' isn't implemented yet).