Rename fullpage option to insert
We will also implement an option to insert scores as inline image (within
a paragraph). Therefore this isn't a binary/boolean option anymore but
an option "insert" with the possible values 'systems' (default), 'fullpage',
and 'inline'.

Inline currently produces an error, but the other two should behave as
expected (of course existing documents have to be updated).