Report "failed" LilyPond compilation and capture log
Previously only the stdout part of LilyPond's log was visible,
which wasn't helpful for tracking down bugs.

With this commit stdout and stderr are merged to a .log file,
which is deleted after compilation if debug-lilypond is not set

If Lilypond exits with non-zero status, a warning is issued,
pointing to both the log and the generated LilyPond code,
which both are *not* deleted in that case, regardless of the setting
of debug-lilypond.

NOTE: I assume the redirection doesn't work on Windows, but I didn't
find a better solution. It seems there is missing a capability similar
to Python's subprocess module that can interact with the shell.