By default don't pass fonts to LilyPond
When the .tex document uses TeX-fonts (i.e. fonts installed in the
texmf tree and not as system fonts passing the text document's fonts to LilyPond will result in the use of ugly fallback fonts. This is usually the case
when the user creates a simple document without explicit font settings.

When weighing between not activating a cool lyluatex feature by default
or risking ugly unexpected results I think we should opt for the less interesting
default.