Implement "intertext" option (WIP)
When "intertext" is set to a non-empty string this string is printed
between the verbatim code and the score.

By default it simply prints a paragraph followed by a skip, but the appearance
can be modified by renewing the \lyIntertext command.

NOTE: The *value* of this option is very unstable ATM for *local* options:
- Anything except a single word or an empty string will simply be ignored

NOTE: The value is unstable in general:
- single quotes are not possible
- macros are not possible