Respect "% begin verbatim" comments
If LilyPond code contains
% begin verbatim and optionally
% end verbatim

Only the part inside that will be printed verbatim.

NOTE: I'm not 100% sure if that fully matches lilypond-book behaviour
because we *only* search for the exact construct "% begin verbatim"
and not ' "begin verbatim" within a comment" as the lilypond-book docs
state (but I doubt lilypond-book does that additional analysis).