Add 'raw-pdf' option
This option does not return any \includegraphics macros or similar
but the raw PDF name(s) (without extension).

For insert=inline|bare-inline|fullpage the plain filename is returned,
including the tmpdir and the hash (i.e.: self.output of the score).
For insert=systems a comma-separated list is returned, containing
the file names as above but also the horizontal offset in the first element:
    -5pt,tmp-ly/my-hash-1,tmp-ly/my-hash-5
(when print-only is set to {1,5} and the left protrusion is 5pt)