Factor out Score:clean_failed_compilation()
In Score:process() we don't want to check for a compilation
*without errors* but only if a score is compiled at all.

Incidentally this reveals that Score:is_compiled_without_errors()
has been conceived somewhat wrong: From its name it should
return true when a score is compiled and LilyPond doesn't report
errors and true when no score is compiled or LilyPond reports an error.

But actually I don't think we ever need exactly this boolean value.
When this function is called from Score:write_latex() we actually don't
want *that*. The condition for the line returning out of the function without writing any LaTeX is Score:is_compiled() but we need the check for the
errorless compilation for reporting and erroring out.
So I'll change is_compiled_without_errors to check_compilation(). This
will return true when Score:is_compiled() returns true, performs the
checks and warnings and calls cleanup_failed_compilation().