Retry lilypond command when failed
Fix #204: MiKTeX behavior isn't the same as TexLive's with io.popen. I
don't know why, but on MiKTeX, we have to add a level of quotes, and
it doesn't seem to be able to write code to stdin.

Incidentally, when LilyPond fails on a score, compilation is re-tried
with debug=true.